+++ src/LegacyAdapter.js
@@ -60,14 +60,16 @@ module.exports = class LegacyAdapter {
 			const outputFile = tmpDirectory + "dist" + path.sep + setName + ".js";
 			const entryFileReal = path.resolve(context.dir.current + config.entry);
 			const entryFileRelative = path.relative(tmpDirectory, entryFileReal);
-			fs.writeFileSync(entryFile, "import \"" + entryFileRelative.replace(/\\/g, "/") + "\";");
+			fs.writeFileSync(entryFile, "module.exports = require( \"" + entryFileRelative.replace(/\\/g, "/") + "\");");
 			apps.push({
 				"entry": path.relative(context.dir.current, entryFile),
 				"output": path.relative(context.dir.current, outputFile),
+				"warningIgnorePattern": config.warningIgnorePattern,
+				"webpackConfig": config.webpackConfig,
 				"@setName": setName,
 				"@isStyle": isStyle,
 				"@legacy": config
-			})
+			});
 		};
 		if (hasCss) context.laborConfig.css.forEach(config => assetToAppConverter(config, true));
 		if (hasJs) context.laborConfig.js.forEach(config => assetToAppConverter(config, false));